refactor: layer and rename src modules
