+++ AGENTS.md
@@ -18,7 +18,7 @@ Treat **AGENTS.md** as the agent-facing index for [README.md](README.md), [docs/
 | Community expectations                     | [docs/CODE_OF_CONDUCT.md](docs/CODE_OF_CONDUCT.md)                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                   |
 | Published CLI help text                    | [assets/help.md](assets/help.md)                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     |
 | CI and security workflows                  | [.github/workflows/quality.yml](.github/workflows/quality.yml), [.github/workflows/verification.yml](.github/workflows/verification.yml), [.github/workflows/security.yml](.github/workflows/security.yml), [.github/workflows/integration.yml](.github/workflows/integration.yml), [.github/workflows/performance.yml](.github/workflows/performance.yml), [.github/workflows/documentation-sync.yml](.github/workflows/documentation-sync.yml), [.github/workflows/anti-slop.yml](.github/workflows/anti-slop.yml), [.github/workflows/publish.yml](.github/workflows/publish.yml) |
-| Library and CLI implementation             | [src/index.ts](src/index.ts), [src/utils.ts](src/utils.ts), [src/bin.ts](src/bin.ts)                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 |
+| Library and CLI implementation             | [src/index.ts](src/index.ts), [src/core/orchestrator.ts](src/core/orchestrator.ts), [src/shared/utils.ts](src/shared/utils.ts), [src/bin.ts](src/bin.ts)                                                                                                                                                                                                                                                                                                                                                                                                                             |
 | tsdown build                               | [tsdown.config.ts](tsdown.config.ts)                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 |
 | Tests and Vitest settings                  | [test/](test/), [vitest.config.ts](vitest.config.ts)                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 |
 | npm scripts and package metadata           | [package.json](package.json)                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         |
